Add files via upload
PlotterWriter Readme

This solution creates a DLL that exposes a CLI interface which wraps code written to communicate with legacy devices attached via USB-to-serial and USB-to-parallel cables.  It has been developed for communicating with a Houston Instruments Image Maker graphic pen plotter, compatible with the better-known HP7475A plotter.  The code is fully functional and supports each HPGL command and instruction that the Image Maker supports, including code to read a series of digitized points and create a series of HPGL instructions to plot those same coordinates.

The DLL is written in both native C++ and C++/CLI.  The native C++ makes all the Win32 calls and does all the work of communicating with the plotter, while the C++/CLI code wraps the native C++ code and exposes a CLI interface that can be called by .net languages.  In this case the test app is written in C#.

The project PlotterDriver contains all the C++ code used in the solution.  It does all the work of locatin